Refactor

+++ tests/acceptance/control/gameservice.rs
@@ -4,13 +4,13 @@ use util::Process;
 
 
 pub struct GameService {
-	pub port   : uint,
+	pub port   : u16,
 	pub process: Process
 }
 
 impl GameService {
 	pub fn start() -> GameService {
-		let port = rand::random::<uint>() % 10000 + 40000;
+		let port = rand::random::<u16>() % 10000 + 40000;
 
 		let mut process = Process::start(
 			"output/bin/vndf-game-service",
